app/webserver/requirements.txt
grpcio==1.62.1
protobuf==4.25.3
Flask==3.0.2